fix: testing structure fixed & forms tests written
